+++ R/digital-ocean.R
@@ -11,6 +11,7 @@ checkAnalogSea <- function() {
 
   suggests <- read.dcf(system.file("DESCRIPTION", package = "plumber"))[1, "Suggests"]
   pkgs <- strsplit(suggests, ",")[[1]]
+  pkgs <- trimws(pkgs)
   analogsea_version <- gsub("[^.0-9]", "", pkgs[grepl("^analogsea ", pkgs)])
   if (utils::packageVersion("analogsea") < package_version(analogsea_version)) {
     stop("The analogsea package is not high enough. Please update `analogsea`.",
@@ -150,7 +151,7 @@ install_nginx <- function(droplet){
 }
 
 install_new_r <- function(droplet){
-  analogsea::droplet_ssh(droplet, "apt-key adv --keyserver keyserver.ubuntu.com --recv-keys E084DAB9")
+  analogsea::droplet_ssh(droplet, "apt-key adv --keyserver keyserver.ubuntu.com --recv-keys 51716619E084DAB9")
   analogsea::droplet_ssh(droplet, "echo 'deb https://cran.rstudio.com/bin/linux/ubuntu focal-cran40/' >> /etc/apt/sources.list.d/cran.list")
   # TODO: use the analogsea version once https://github.com/sckott/analogsea/issues/139 is resolved
   #analogsea::debian_apt_get_update(droplet)