+++ R/pr.R
@@ -342,6 +342,11 @@ pr_hooks <- function(pr,
 #'   Defaults to \code{TRUE}.
 #' @param secure Boolean that adds the \code{Secure} cookie flag.  This should be set
 #'   when the route is eventually delivered over \href{https://en.wikipedia.org/wiki/HTTPS}{HTTPS}.
+#' @param sameSite A character specifying the SameSite policy to attach to the cookie.
+#'   If specified, one of the following values should be given: "Strict", "Lax", or "None".
+#'   If "None" is specified, then the \code{secure} flag MUST also be set for the modern browsers to
+#'   accept the cookie. An error will be returned if \code{sameSite = "None"} and \code{secure = FALSE}.
+#'   If not specified or a non-character is given, no SameSite policy is attached to the cookie.
 #' @seealso \itemize{
 #' \item \href{https://github.com/jeroen/sodium}{'sodium'}: R bindings to 'libsodium'
 #' \item \href{https://download.libsodium.org/doc/}{'libsodium'}: A Modern and Easy-to-Use Crypto Library
@@ -402,10 +407,11 @@ pr_cookie <- function(pr,
                       name = "plumber",
                       expiration = FALSE,
                       http = TRUE,
-                      secure = FALSE) {
+                      secure = FALSE,
+                      sameSite = FALSE) {
   validate_pr(pr)
   pr$registerHooks(
-    sessionCookie(key = key, name = name, expiration = expiration, http = http, secure = secure)
+    sessionCookie(key = key, name = name, expiration = expiration, http = http, secure = secure, sameSite = sameSite)
   )
   invisible(pr)
 }