+++ R/session-cookie.R
@@ -40,6 +40,11 @@
 #'   Defaults to \code{TRUE}.
 #' @param secure Boolean that adds the \code{Secure} cookie flag.  This should be set
 #'   when the route is eventually delivered over \href{https://en.wikipedia.org/wiki/HTTPS}{HTTPS}.
+#' @param sameSite A character specifying the SameSite policy to attach to the cookie.
+#'   If specified, one of the following values should be given: "Strict", "Lax", or "None".
+#'   If "None" is specified, then the \code{secure} flag MUST also be set for the modern browsers to
+#'   accept the cookie. An error will be returned if \code{sameSite = "None"} and \code{secure = FALSE}.
+#'   If not specified or a non-character is given, no SameSite policy is attached to the cookie.
 #' @export
 #' @seealso \itemize{
 #' \item \href{https://github.com/jeroen/sodium}{'sodium'}: R bindings to 'libsodium'
@@ -95,7 +100,8 @@ sessionCookie <- function(
   name = "plumber",
   expiration = FALSE,
   http = TRUE,
-  secure = FALSE
+  secure = FALSE,
+  sameSite = FALSE
 ) {
 
   if (missing(key)) {
@@ -104,7 +110,19 @@ sessionCookie <- function(
   key <- asCookieKey(key)
 
   # force the args to evaluate
-  list(expiration, http, secure)
+  list(expiration, http, secure, sameSite)
+
+  # sanity check the sameSite and secure arguments
+  if (is.character(sameSite)) {
+    sameSite <- match.arg(sameSite, c("Strict", "Lax", "None"))
+  } else {
+    sameSite <- FALSE
+  }
+  if (identical(sameSite, "None")) {
+    if (!secure) {
+      stop("You must set `secure = TRUE` when `sameSite = \"None\"`. See: https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Set-Cookie")
+    }
+  }
 
   # Return a list that can be added to registerHooks()
   list(
@@ -129,13 +147,13 @@ sessionCookie <- function(
       session <- req$session
       # save session in a cookie
       if (!is.null(session)) {
-        res$setCookie(name, encodeCookie(session, key), expiration = expiration, http = http, secure = secure)
+        res$setCookie(name, encodeCookie(session, key), expiration = expiration, http = http, secure = secure, sameSite = sameSite)
       } else {
         # session is null
         if (!is.null(req$cookies[[name]])) {
           # no session to save, but had session to parse
           # remove cookie session cookie
-          res$removeCookie(name, "", expiration = expiration, http = http, secure = secure)
+          res$removeCookie(name, "", expiration = expiration, http = http, secure = secure, sameSite = sameSite)
         }
       }