+++ tests/testthat/test-injection.R
@@ -3,6 +3,5 @@ context("Injection")
 test_that("Injected arguments on req$args get passed on.", {
   r <- plumber$new(test_path("files/filter-inject.R"))
 
-  res <- PlumberResponse$new()
-  expect_equal(r$serve(make_req("GET", "/"), res)$body, jsonlite::toJSON(13))
+  expect_equal(r$call(make_req("GET", "/"))$body, jsonlite::toJSON(13))
 })